requirements.txt
streamlit
pandas
numpy
plotly
streamlit-option-menu

streamlit
pandas
pymupdf
openai>=1.0.0
python-dotenv
plotly
matplotlib
seaborn

langchain-core
langchain-community
langchain-openai
langchain-experimental
langchain-text-splitters

faiss-cpu

pillow

typing-extensions
tqdm

pandas
numpy
matplotlib
seaborn
chainlit
Pillow
google-generativeai
python-dotenv
nest-asyncio